+++ glide-core/redis-rs/redis/src/cmd.rs
@@ -6,7 +6,7 @@ use futures_util::{
 };
 #[cfg(feature = "aio")]
 use std::pin::Pin;
-use std::{borrow::Borrow, fmt, io};
+use std::{borrow::Borrow, fmt, io, sync::Arc};
 
 use crate::connection::ConnectionLike;
 use crate::pipeline::Pipeline;
@@ -371,8 +371,11 @@ impl Cmd {
     ///
     /// A span is used by an OpenTelemetry backend to track the lifetime of the command
     #[inline]
-    pub fn with_span(&mut self, name: &str) -> &mut Cmd {
-        self.span = Some(telemetrylib::GlideOpenTelemetry::new_span(name));
+    pub fn with_span_by_ptr(&mut self, span_ptr: u64) -> &mut Cmd {
+        unsafe {
+            Arc::increment_strong_count(span_ptr as *const GlideSpan);
+            self.span = Some((*unsafe { Arc::from_raw(span_ptr as *const GlideSpan) }).clone());
+        }
         self
     }